+++ src/SidePanel/SidePanel.tsx
@@ -38,17 +38,19 @@ import React from 'react';
 import { SidePanel } from 'pc-nrfconnect-shared';
 
 import AdvancedOptions from './AdvancedOptions';
+import Instructions from './Instructions';
 import TraceCollector from './TraceCollector';
 import TraceConverter from './TraceConverter';
-import Wireshark from './Wireshark';
+import TracefileInformation from './TracefileInformation';
 
 import './sidepanel.scss';
 
 export default () => (
     <SidePanel className="side-panel">
         <TraceCollector />
+        <Instructions />
         <TraceConverter />
-        <Wireshark />
+        <TracefileInformation />
         <AdvancedOptions />
     </SidePanel>
 );