+++ src/SidePanel/TraceConverter.tsx
@@ -1,11 +1,18 @@
 import React from 'react';
 import Button from 'react-bootstrap/Button';
-import { useDispatch } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 
 import { convertTraceFile } from '../nrfml/nrfml';
+import { getSerialPort } from '../reducer';
 import { askForTraceFile } from '../utils/fileUtils';
 
 export default () => {
+    const selectedSerialPort = useSelector(getSerialPort);
+
+    if (selectedSerialPort) {
+        return null;
+    }
+
     const dispatch = useDispatch();
 
     const loadTrace = () => {
@@ -21,7 +28,7 @@ export default () => {
             variant="secondary"
             onClick={loadTrace}
         >
-            Convert Trace
+            Convert Raw Trace to PCAP
         </Button>
     );
 };